requirements.txt
#
# This file is autogenerated by pip-compile
# Make changes in requirements.in, then run this to update:
#
#    pip-compile requirements.in
#
amqp==1.4.7
anyjson==0.3.3
apipkg==1.4               # via execnet
billiard==3.3.0.20
celery==3.1.18
click==5.1                # via python-dotenv
coverage==4.0             # via pytest-cov
decorator==4.0.4          # via ipython, traitlets
defusedxml==0.4.1         # via python3-openid
django-appconf==1.0.1
django-bootstrap3==6.2.2
django-celery==3.1.16
django-compressor==1.5
django-datetime-widget==0.9.3
django-debug-toolbar==1.3.2
django-extensions==1.5.7
django-mailgun==0.7.2
django-phonenumber-field==0.7.2
django-solo==1.1.0
django-twilio==0.8.0
django==1.8.4
djangorestframework==3.2.4
elvantoapi==1.3.3.2
execnet==1.4.1            # via pytest-cache, pytest-xdist
gunicorn==19.3.0
httplib2==0.9.2           # via twilio
ipython-genutils==0.1.0   # via traitlets
ipython==4.0.0
isort==4.2.2              # via pytest-isort
kombu==3.0.26             # via celery
oauthlib==1.0.3           # via python-social-auth, requests-oauthlib
opbeat==3.1.1
path.py==8.1.1            # via pickleshare
pep8==1.6.2               # via pytest-pep8
pexpect==3.3              # via ipython
phonenumbers==7.0.11      # via django-phonenumber-field
pickleshare==0.5          # via ipython
psycopg2==2.6.1
py==1.4.30                # via pytest, pytest-xdist
pyflakes==1.0.0           # via pytest-flakes
pygal==2.0.7
pyjwt==1.4.0              # via python-social-auth
pysocks==1.5.6            # via twilio
pytest-cache==1.0         # via pytest-flakes, pytest-isort, pytest-pep8
pytest-cov==2.1.0
pytest-django==2.8.0
pytest-env==0.5.1
pytest-flakes==1.0.1
pytest-isort==0.1.0
pytest-pep8==1.0.6
pytest-xdist==1.13.1
pytest==2.8.1
python-dotenv==0.1.3
python-social-auth==0.2.13
python3-memcached==1.51
python3-openid==3.0.7     # via python-social-auth
pytz==2015.6              # via celery, django-datetime-widget, twilio
requests-oauthlib==0.5.0  # via python-social-auth
requests==2.7.0
simplegeneric==0.8.1      # via ipython
six==1.9.0                # via django-appconf, django-extensions, django-mailgun, python-social-auth, twilio
sqlparse==0.1.16          # via django-debug-toolbar
traitlets==4.0.0          # via ipython
twilio==4.6.0
